feat: custom docs
As we now have custom constructors, we need a way to include them in the
docs as well. This adds a pretty basic, yet usable way:

Overriding the internal data model using Jsonnet. This is done in
`.docsonnet` files, which are automatically added upon what we generate
from swagger.json